feat(react css modules): including creation of CSS module during the creation of the React component
During the steps of the wizard, once React is selected as the framework, you will be prompted to
choose a CSS framework. In this list of options, the 'CSS Module' option has been added. If
selected, a new template is used for component generation, and the CSS module is generated and
imported into the new component.

closes #26